set mbed target and toolchain every time
some users report that the mbed-target and mbed-toolchain
are not being set automatically during the first invocation
of 'make', causing the build the fail.

if we set the values on every build, then .mbed should always
have the values specified.

Note the Makefile checks the current value of these parameters
and sets it to the same value if it already exists so that
user-specified config is not overwritten.