set mbed target and toolchain every time
if mbed-import is used to clone the ref-wem-firmware project
instead of git-clone, then a .mbed file already exists in the
root project directory before the 'make' command is invoked
the first time.  this prevented the Makefile from setting
the mbed-target and mbed-toolchain parameters which resulted
in failed builds because the values were not specified.

if we set the values on every build, then .mbed should always
have the values specified.

Note the Makefile checks the current value of these parameters
and sets it to the same value if it already exists so that
user-specified config is not overwritten.